Separates TicTacToe class into Board and DisplayBoard classes
