+++ tic_tac_toe/__main__.py
@@ -1,0 +1,3 @@
+from tic_tac_toe.game import run
+
+run()
