+++ tic_tac_toe/game.py
@@ -1,0 +1,8 @@
+from tic_tac_toe.board import Board
+from tic_tac_toe.views import BoardDisplay
+
+
+def run():
+    board = Board()
+    view = BoardDisplay()
+    view.display_board(board)
