--- tictactoe.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-from tic_tac_toe.tic_tac_toe import TicTacToe
-
-
-if __name__ == "__main__":
-    game = TicTacToe([" ", " ", " ", " ", " ", " ", " ", " ", " "])
-
-    game.play()
